cleanup and unfication of new liveliness tests
