--- tests/DCPS/LivelinessLocal/LivelinessTest.mpc
@@ -1,14 +1,0 @@
-project(LivelinessLocal): dcpsexe, dcps_test, dcps_transports_for_test {
-  exename   = publisher
-  libpaths += ../FooType4
-  libpaths += ../common
-  libs     +=  common DcpsFooType4
-  after    += DcpsFooType4 common
-
-  Source_Files {
-    common.cpp
-    DataReaderListener.cpp
-    publisher.cpp
-    Writer.cpp
-  }
-}
