--- tests/DCPS/LivelinessLocal/README
@@ -1,7 +1,0 @@
-A basic liveliness test of DDS.
-
-run_test.pl
-  Run a separate publisher and subscriber process.
-
-run_test.pl take
-  Run the same test but "take" instead of "read".
