--- tests/DCPS/LivelinessLocal/rtps_disc.ini
@@ -1,14 +1,0 @@
-[common]
-DCPSGlobalTransportConfig=rtps
-DCPSDefaultDiscovery=uni_rtps
-
-[transport/rtps]
-transport_type=rtps_udp
-use_multicast=0
-
-[config/rtps]
-transports=rtps
-
-[rtps_discovery/uni_rtps]
-SedpMulticast=0
-ResendPeriod=2
