--- tests/DCPS/LivelinessLocal/tcp.ini
@@ -1,6 +1,0 @@
-[common]
-DCPSInfoRepo=file://repo.ior
-DCPSGlobalTransportConfig=$file
-
-[transport/tcp]
-transport_type=tcp
