--- tests/DCPS/LivelinessLocal/udp.ini
@@ -1,6 +1,0 @@
-[common]
-DCPSInfoRepo=file://repo.ior
-DCPSGlobalTransportConfig=$file
-
-[transport/udp]
-transport_type=udp
