--- tests/DCPS/LivelinessNotify/.gitignore
@@ -1,2 +1,0 @@
-/publisher
-/subscriber
