--- tests/DCPS/LivelinessNotify/LivelinessTest.mpc
@@ -1,28 +1,0 @@
-project(LivelinessNotifyTestPublisher): dcpsexe, dcps_test, dcps_transports_for_test {
-  exename   = publisher
-  libpaths += ../FooType4
-  libpaths += ../common
-  libs     +=  common DcpsFooType4
-  after    += DcpsFooType4 common
-
-  Source_Files {
-    common.cpp
-    publisher.cpp
-    Writer.cpp
-  }
-}
-
-
-project(LivelinessNotifyTestSubscriber): dcpsexe, dcps_test, dcps_transports_for_test {
-  exename   = subscriber
-  libpaths += ../FooType4
-  libpaths += ../common
-  libs     +=  common DcpsFooType4
-  after    += DcpsFooType4 common
-
-  Source_Files {
-    common.cpp
-    subscriber.cpp
-    DataReaderListener.cpp
-  }
-}
