--- tests/DCPS/LivelinessNotify/README
@@ -1,18 +1,0 @@
-Currently, the OpenDDS Liveliness does not account for automatic or manual liveliness,
-and will instead notify all subscribers when the participant receives a liveliness
-notification.  This test will create 2 participants. One participant will contain 2
-writers, one which is automatic liveliness and the other which is manual liveliness.
-The other participant will contain 2 readers, one which is automatic liveliness and
-the other which is manual liveliness. The automatic writer will only associate with
-the automatic reader, and the manual writer will associate with the manual reader. The
-automatic reader will send a liveliness message letting the automatic reader know it
-is alive.  The currently implementation of OpenDDS will incorrectly let the manual
-reader know that the participant received a liveliness message, and the manual reader
-will incorrectly assume that the manual writer is still alive, despite not receiving
-a liveliness message in the time limit.
-
-run_test.pl
-  Run a separate publisher and subscriber process.
-
-run_test.pl take
-  Run the same test but "take" instead of "read".
