+++ tests/DCPS/LivelinessTest/.gitignore
@@ -1,2 +1,5 @@
 /publisher
 /subscriber
+/dualpub
+/dualsub
+/local
